Bug 1134028 - Don't assign flags to file that may not be defined
The renderer/d3d/d3d11/SwapChain11.cpp source file is only defined if
building on Windows. The generated moz.build was adjusting the .flags
of this file unconditionally. This worked up until now because the
moz.build was only being evaluated on Windows. However, this assumption
no longer holds true after work in bug 1132771 is complete.

Patch the generation script to put .flags adjustment for the file
in question behind the same conditional as the SOURCES assignment.